+++ fast_rx_test/lib/src/rx_test.dart
@@ -4,6 +4,12 @@ import 'package:fast_rx/src/rx/rx.dart';
 
 /// A container for a [transform] on [rx]
 class RxTest<T extends Rx> {
+  /// An optional id of the test for debugging
+  final String? id;
+
+  /// If true, print notification values to the console
+  final bool verbose;
+
   /// The rx value
   final T rx;
 
@@ -11,5 +17,5 @@ class RxTest<T extends Rx> {
   final void Function(T rx) transform;
 
   /// Create an [RxTest]
-  RxTest(this.rx, this.transform);
+  RxTest(this.rx, this.transform, {this.id, this.verbose = false});
 }